Update router to pass props
