+++ src/components/Login.jsx
@@ -9,7 +9,6 @@ const LogIn = (props) => {
 
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
-    //const dispatch= useDispatch();
     const dispatch= useDispatch();
 
     function onRegister() {
